test: 修改react-native-cookies-cookies测试用例
